Moved repeated config methods for tests to a single file
