+++ tests/unit_tests/test_inference.py
@@ -12,19 +12,10 @@ from core.ports.database_port import DatabasePort
 from tests.mocks.authentication_mock import AuthenticationMock
 from tests.mocks.mongo_mock import MongoMock
 
-
-def configure_ports_without_auth():
-    ports = {}
-    ports["database_port"] = DatabasePort(MongoMock())
-    ports["authentication_port"] = AuthenticationPort(AuthenticationAdapter())
-    return ports
-
-
-def configure_ports_with_auth():
-    ports = {}
-    ports["database_port"] = DatabasePort(MongoMock())
-    ports["authentication_port"] = AuthenticationPort(AuthenticationMock())
-    return ports
+from tests.unit_tests.config import (
+    configure_ports_without_auth,
+    configure_ports_with_auth,
+)
 
 
 @pytest.fixture()
