+++ tests/unit_tests/test_user.py
@@ -14,24 +14,10 @@ from tests.mocks.authentication_mock import AuthenticationMock
 from tests.mocks.mongo_mock import MongoMock
 
 
-def configure_ports_without_auth():
-    ports = {}
-    ports["database_port"] = DatabasePort(MongoMock())
-    ports["authentication_port"] = AuthenticationPort(AuthenticationAdapter())
-    return ports
-
-
-def configure_ports_with_auth():
-    ports = {}
-    ports["database_port"] = DatabasePort(MongoMock())
-    ports["authentication_port"] = AuthenticationPort(AuthenticationMock())
-    return ports
-
-
-def plug_adapters_to_ports():
-    ports = {}
-    ports["database"] = DatabasePort(MongoMock())
-    return ports
+from tests.unit_tests.config import (
+    configure_ports_without_auth,
+    configure_ports_with_auth,
+)
 
 
 @pytest.fixture()
